Refactored + fixed tests + fix flow issues
