--- GreetingService.js
@@ -1,20 +1,0 @@
-// @flow
-
-
-class GreetingService {
-  hello(name) {
-    return `Hello ${name}`;
-  }
-}
-GreetingService.meta = {
-  type: 'class',
-  methods: {
-    hello: {
-      type: 'sync'
-    }
-  }
-}
-;
-
-
-export default GreetingService;
