+++ examples/GreetingServiceClass/App.js
@@ -1,7 +1,7 @@
 // @flow
 //import * as ServiceA from './ServiceA.js';
 import GreetingService from './GreetingService.js';
-import { Microservices } from './sdk/Microservices.js';
+import { Microservices } from 'src/scalecube-services';
 
 
 const greetingService = Microservices.builder().services(GreetingService, GreetingService).build().proxy().api(GreetingService);
