--- old/ServiceA.js
@@ -1,7 +1,0 @@
-// @flow
-
-export const getUser = () => "idan";
-getUser.isService = true;
-getUser.serviceMeta = {
-  type: 'promise'
-};
