--- old/dispatcher.js
@@ -1,6 +1,0 @@
-// @flow
-
-interface dispatcher {
-  invoke (message: Message): Future<Message>;
-  listen (message: Message): Subscriber<Message>;
-}
