+++ src/scalecube-services/RoundRobinRouter.js
@@ -1,0 +1,25 @@
+// @flow
+
+export class RoundRobinServiceRouter implements Router {
+  registry: ServiceRegistery;
+  counter: Set<string, number>;
+  constructor(registry: ServiceRegistery) {
+    this.registry = registry;
+    this.counter = new Map();
+  }
+  route(request:Messsage): ServiceInstance {
+    const instances = this.registry.serviceLookup(request.serviceName)
+      .filter(inst => inst.serviceDefinition.methods[request.method] !== "undefined" );
+
+    if( instances.length > 1 ) {
+      let index = (this.counter.get(request.serviceName) || 0) + 1 % instances.length;
+      this.counter.set(request.serviceName, index);
+      return instances[index];
+    } else if ( instances.length == 1 ) {
+      return instances[0];
+    } else {
+      return null;
+    }
+
+  }
+}
