+++ src/scalecube-services/Router.js
@@ -1,0 +1,4 @@
+// @flow
+export interface Router {
+  route(message:Message): ServiceInstance;
+}
