+++ src/scalecube-services/ServiceCall.js
@@ -1,0 +1,12 @@
+// @flow
+
+export class ServiceCall{
+  router;
+  constructor(router){
+    this.router = router;
+  }
+  invoke(message) {
+    const inst = this.router.route(message).service;
+    return inst[message.method](...message.data);
+  }
+}
