+++ src/scalecube-services/index.js
@@ -1,0 +1,26 @@
+// @flow
+
+import { DispatcherContext } from './DispatcherContext';
+import { Microservices, MicroservicesBuilder } from './Microservices';
+import { ProxyContext } from './ProxyContext';
+import { RoundRobinServiceRouter } from './RoundRobinRouter';
+import { Router } from './Router';
+import { ServiceCall } from './ServiceCall';
+import { ServicesConfig } from './ServicesConfig';
+import { ServiceDefinition } from './ServiceDefinition';
+import { ServiceRegistery } from './ServiceRegistery';
+import * as utils from './utils';
+
+export {
+  DispatcherContext,
+  Microservices,
+  MicroservicesBuilder,
+  ProxyContext,
+  RoundRobinServiceRouter,
+  Router,
+  ServiceCall,
+  ServicesConfig,
+  ServiceDefinition,
+  ServiceRegistery,
+  utils
+};
